es-service/requirements.txt
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
elastic-transport==8.4.0
elasticsearch==8.6.0
Flask==2.2.2
Flask-Cors==3.0.10
flask-oidc==1.4.0
Flask-SQLAlchemy==3.0.2
flaskoidc==1.1.2
httplib2==0.21.0
idna==3.4
itsdangerous==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.46
twilio==7.16.1
urllib3==1.26.14
Werkzeug==2.2.2
boto3
python-dotenv
uWSGI